argc: sema: sepeprate module
